Revert "ref(node): Move non-handler code out of `handlers` module (#5190)"
This reverts commit 7c521e3cc4a6a8caa685e2f1316a133b3acae212.
