+++ packages/node/src/index.ts
@@ -15,7 +15,6 @@ export type {
   Thread,
   User,
 } from '@sentry/types';
-export type { AddRequestDataToEventOptions, ExpressRequest } from '@sentry/utils';
 
 export type { NodeOptions } from './types';
 
@@ -42,7 +41,6 @@ export {
   setUser,
   withScope,
 } from '@sentry/core';
-export { addRequestDataToEvent, extractRequestData } from '@sentry/utils';
 
 export { NodeClient } from './client';
 export { makeNodeTransport } from './transports';
