+++ packages/serverless/src/gcpfunction/http.ts
@@ -1,29 +1,17 @@
-import {
-  addRequestDataToEvent,
-  AddRequestDataToEventOptions,
-  captureException,
-  flush,
-  getCurrentHub,
-  startTransaction,
-} from '@sentry/node';
+import { captureException, flush, getCurrentHub, Handlers, startTransaction } from '@sentry/node';
 import { extractTraceparentData } from '@sentry/tracing';
 import { isString, logger, parseBaggageString, stripUrlQueryAndFragment } from '@sentry/utils';
 
 import { domainify, getActiveDomain, proxyFunction } from './../utils';
 import { HttpFunction, WrapperOptions } from './general';
 
-// TODO (v8 / #5190): Remove this
-interface OldHttpFunctionWrapperOptions extends WrapperOptions {
-  /**
-   * @deprecated Use `addRequestDataToEventOptions` instead.
-   */
-  parseRequestOptions: AddRequestDataToEventOptions;
-}
-interface NewHttpFunctionWrapperOptions extends WrapperOptions {
-  addRequestDataToEventOptions: AddRequestDataToEventOptions;
+type ParseRequestOptions = Handlers.ParseRequestOptions;
+
+export interface HttpFunctionWrapperOptions extends WrapperOptions {
+  parseRequestOptions: ParseRequestOptions;
 }
 
-export type HttpFunctionWrapperOptions = OldHttpFunctionWrapperOptions | NewHttpFunctionWrapperOptions;
+const { parseRequest } = Handlers;
 
 /**
  * Wraps an HTTP function handler adding it error capture and tracing capabilities.
@@ -52,13 +40,9 @@ export function wrapHttpFunction(
 
 /** */
 function _wrapHttpFunction(fn: HttpFunction, wrapOptions: Partial<HttpFunctionWrapperOptions> = {}): HttpFunction {
-  // TODO (v8 / #5190): Remove this
-  // eslint-disable-next-line deprecation/deprecation
-  const { parseRequestOptions } = wrapOptions as OldHttpFunctionWrapperOptions;
-
   const options: HttpFunctionWrapperOptions = {
     flushTimeout: 2000,
-    addRequestDataToEventOptions: parseRequestOptions ? parseRequestOptions : {},
+    parseRequestOptions: {},
     ...wrapOptions,
   };
   return (req, res) => {
@@ -88,7 +72,7 @@ function _wrapHttpFunction(fn: HttpFunction, wrapOptions: Partial<HttpFunctionWr
     // since functions-framework creates a domain for each incoming request.
     // So adding of event processors every time should not lead to memory bloat.
     getCurrentHub().configureScope(scope => {
-      scope.addEventProcessor(event => addRequestDataToEvent(event, req, options.addRequestDataToEventOptions));
+      scope.addEventProcessor(event => parseRequest(event, req, options.parseRequestOptions));
       // We put the transaction on the scope so users can attach children to it
       scope.setSpan(transaction);
     });
