+++ packages/utils/package.json
@@ -24,9 +24,6 @@
     "array.prototype.flat": "^1.3.0",
     "chai": "^4.1.2"
   },
-  "peerDependencies": {
-    "cookie": "^0.4.1"
-  },
   "scripts": {
     "build": "run-p build:rollup build:types",
     "build:dev": "run-s build",
